+++ CHANGELOG.md
@@ -1,13 +1,29 @@
 # Change Log
 
-## v2.7-RC-dev2
+## v2.7-RC-dev3
+
++ 优化cmake组合方式(所以推荐一并更新全部的Cmake文件)
++ loader增加jline3命令输入框
+ - 支持运行时登录
+ - 支持调用onDisable退出
++ loader支持md5作为密码(32位md5 大小写都可以)
++ 更新mirai版本到v2.7-RC-dev1 修复管理员识别问题
++ 支持群公告操作
++ 支持上下文
++ 添加ServiceMessage支持(UrlSharer)
++ 修复一些MiraiCode的小问题
++ 修复Mirai端取消mute(0)导致的不能unMute
+
+> 该版本为预览版的开发版(，包含一些可能出现问题的特性
+
+## v2.7-RC-dev2 [deleted and merge into dev3]
 
 + 修复Mirai端取消mute(0)导致的不能unMute
 + 修复loader一开始启动没有颜色
 
 > 该版本为v2.7-RC-dev的补丁
 
-## v2.7-RC-dev
+## v2.7-RC-dev [deleted and merge into dev3]
 
 + 优化cmake组合方式(所以推荐一并更新全部的Cmake文件)
 + loader增加jline3命令输入框